Replace _Sequence by _FakeSequence
The auto-stretch feature of the _Sequence contained hidden magic,
as it tracked the last "entry" that has been accessed and pretended
to have a length based on this. So, outcomes depend on something that
is assumed to be a read access. Also, self.data was erived from UserList
but never really supported all of the expected functionality in a
consistent way.

The new _FakeSequence is a lot simpler. In particular it has an explicit
(fake) length that needs to be set before casting to (stretched) lists.
The default length is still set, because there is an entry, and to allow
to directly iterate over a _FakeSequence, e.g. in sum().