Allow filtering the result `DataFrame`'s columns
When calling `results.to_df(variable)`, look for a `filters` attribute
on `results`, then look up `results.filters[variable]`, which should be
a function, and use that function to filter the `columns` of the
`variable`'s results `DataFrame` before returning it. More specifically,
the function should accept entries of the `DataFrame`'s `columns`
attribute and return something that evaluates to `True` if the column
should be included in the results and to `False` if the column should be
excluded.
Retain all columns, if `results.filters` does not contain the key
`variable` and provide a `filters` class attribute on `Results` in order
to provide default filters.
Add a `filters` argument to the `Results` constructor to allow setting
the `filters` attribute on construction.