Make result filters easily customizable
Create as special `dict` subclass for `Filters` for `Results`. The
subclass as a special `updating` method, which is functionally
equivalent to `dict`'s `update` method, except that it doesn't update in
place, but creates a shallow clone and returns the `update`d clone.
The idea is that this enables combining the default `Results.filters`
easily with custom ones by doing e.g.:

```python
def nosinks(column):
    return not isinstance(column[1], Sink)

results.filters = Results.filters.updating({"flow": nosinks})
flows = results["flow"]
```

Note that the code making `Results.filters` a `Filters` instance instead
of a regular `dict` is not properly formatted yet. This is intentional,
in order to make it easier to spot that the code didn't change
semantically beyond adding the `Filters` call. Formatting will be done
in the next commit.