Test `hasattr` for `Results` instances
Test that `hasattr` works correctly when checking whether an attribute
is present on a `Results` instance. The `Results.__getattr__`
implementation currently has a bug that results in a `KeyError` when
checking for an attribute that isn't present. This is because the
implementation does not raise an `AttributeError` as specified by the
[documentation][0], but instead always tries item access, which in turn
calls `.to_df`, and passes any resulting errors unmodified.
This problem is also mentioned in issue #1237.

[0]: https://docs.python.org/3.15/reference/datamodel.html#object.__getattr__