Remove attribute access to optimization `Results`
Streamline the API by removing attribute access to optimization
`Results`. This means the attributes of a `Results` object are only
there in order to help with processing and the actual optimization
results can only be accessed via item access. This separates the
concerns more clearly and is thus less confusing for the user.

Adjust the tests accordingly. The `"flow"` and `"status"` variables are
now exclusively accessible via item access.